requirements.txt
langchain_google_genai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
groq
langchain_groq
beautifulsoup4
langchainhub
faiss-cpu
cassio
arxiv
wikipedia
chromadb
pypdf
langchain-classic